refactor: t193 extract AgentLoop subresponsibilities into focused classes
AgentLoop was ~1870 lines handling 8+ concerns. Extract into focused classes:

- SystemInstructionBuilder: build_system_instruction() + memory/skill/knowledge assembly
- ProviderCredentialLoader: ensure_provider_credentials_static()
- ToolPermissionResolver: get_tools_needing_confirmation(), classify_ability(),
  set_always_allow(), get_always_allowed()
- SpinDetector: build_tool_signature(), idle-round tracking state
- ClientAbilityRouter: partition_tool_calls(), client ability stubs, name resolution
- ConversationSerializer: serialize/deserialize history, append_tool_response, truncate

AgentLoop becomes a ~1190-line orchestrator that delegates to these services.
All existing public static methods are preserved as backward-compatible stubs.

Resolves #995
